Fix concurrency issue
