+++ http/org.wso2.carbon.transport.http.netty/src/main/java/org/wso2/carbon/transport/http/netty/message/MessageFuture.java
@@ -2,6 +2,8 @@ package org.wso2.carbon.transport.http.netty.message;
 
 import io.netty.handler.codec.http.HttpContent;
 
+import java.util.concurrent.ConcurrentLinkedQueue;
+
 /**
  * Represents future contents of the message
  */
@@ -9,29 +11,41 @@ public class MessageFuture {
 
     private MessageListener messageListener;
     private HTTPCarbonMessage httpCarbonMessage;
+    private ConcurrentLinkedQueue<HttpContent> pendingPayload;
 
-    public MessageFuture(HTTPCarbonMessage httpCarbonMessage) {
+     public MessageFuture(HTTPCarbonMessage httpCarbonMessage) {
         this.httpCarbonMessage = httpCarbonMessage;
+        this.pendingPayload = new ConcurrentLinkedQueue<>();
+//        while (!httpCarbonMessage.isEmpty()) {
+//            HttpContent httpContent = httpCarbonMessage.getHttpContent();
+//            pendingPayload.add(httpContent);
+//        }
     }
 
-    public void setMessageListener(MessageListener messageListener) {
+    public synchronized void setMessageListener(MessageListener messageListener) {
         this.messageListener = messageListener;
-        while (!httpCarbonMessage.getBlockingEntityCollector().isEmpty()) {
-            notifyMessageListener(httpCarbonMessage.getBlockingEntityCollector().getHttpContent());
+        while (!httpCarbonMessage.isEmpty()) {
+            HttpContent httpContent = httpCarbonMessage.getHttpContent();
+            notifyMessageListener(httpContent);
+        }
+        while (!pendingPayload.isEmpty()) {
+            notifyMessageListener(pendingPayload.poll());
         }
     }
 
-    public void removeMessageListener() {
+    public synchronized void removeMessageListener() {
         this.messageListener = null;
     }
 
-    public void notifyMessageListener(HttpContent httpContent) {
+    public synchronized void notifyMessageListener(HttpContent httpContent) {
         if (this.messageListener != null) {
             this.messageListener.onMessage(httpContent);
+        } else {
+            pendingPayload.add(httpContent);
         }
     }
 
-    public HttpContent sync() {
+    public synchronized HttpContent sync() {
         return this.httpCarbonMessage.getBlockingEntityCollector().getHttpContent();
     }
 }
