+++ http/org.wso2.carbon.transport.http.netty/src/main/java/org/wso2/carbon/transport/http/netty/sender/channel/TargetChannel.java
@@ -150,8 +150,6 @@ public class TargetChannel {
                 HTTPTransportContextHolder.getInstance().getHandlerExecutor().
                         executeAtTargetRequestReceiving(httpCarbonRequest);
             }
-
-            Util.setupTransferEncodingForRequest(httpCarbonRequest, chunkDisabled);
             HttpRequest httpRequest = Util.createHttpRequest(httpCarbonRequest);
 
             this.setRequestWritten(true);
