+++ http/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/contentaware/ContentAwareMessageProcessorTestCase.java
@@ -25,10 +25,11 @@ import org.testng.annotations.Test;
 import org.wso2.carbon.messaging.exceptions.ServerConnectorException;
 import org.wso2.carbon.transport.http.netty.config.TransportsConfiguration;
 import org.wso2.carbon.transport.http.netty.config.YAMLTransportConfigurationBuilder;
-import org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners.RequestResponseCreationListener;
-import org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners.RequestResponseCreationStreamingListener;
-import org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners.RequestResponseTransformListener;
-import org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners.RequestResponseTransformStreamingListener;
+import org.wso2.carbon.transport.http.netty.contentaware.listeners.RequestResponseCreationListener;
+import org.wso2.carbon.transport.http.netty.contentaware.listeners.RequestResponseCreationStreamingListener;
+import org.wso2.carbon.transport.http.netty.contentaware.listeners.RequestResponseTransformListener;
+import org.wso2.carbon.transport.http.netty.contentaware.listeners.RequestResponseTransformStreamingListener;
+import org.wso2.carbon.transport.http.netty.contentaware.listeners.ResponseStreamingWithoutBufferingListener;
 import org.wso2.carbon.transport.http.netty.contract.HttpConnectorListener;
 import org.wso2.carbon.transport.http.netty.contract.ServerConnector;
 import org.wso2.carbon.transport.http.netty.passthrough.PassthroughMessageProcessorListener;
@@ -156,23 +157,23 @@ public class ContentAwareMessageProcessorTestCase {
         }
     }
 
-//    @Test
-//    public void responseStreamingWithoutBufferingTestCase() {
-//        String requestValue = "<A><B><C>Test Message</C></B></A>";
-//        try {
-//            httpConnectorListener = new ResponseStreamingWithoutBufferingListener();
-//            TestUtil.updateMessageProcessor(httpConnectorListener);
-//            HttpURLConnection urlConn = TestUtil.request(baseURI, "/", HttpMethod.POST.name(), true);
-//            urlConn.setChunkedStreamingMode(-1); // Enable Chunking
-//            TestUtil.writeContent(urlConn, requestValue);
-//            assertEquals(200, urlConn.getResponseCode());
-//            String content = TestUtil.getContent(urlConn);
-//            assertEquals(requestValue, content);
-//            urlConn.disconnect();
-//        } catch (IOException e) {
-//            TestUtil.handleException("IOException occurred while running responseStreamingWithoutBufferingTestCase", e);
-//        }
-//    }
+    @Test
+    public void responseStreamingWithoutBufferingTestCase() {
+        String requestValue = "<A><B><C>Test Message</C></B></A>";
+        try {
+            httpConnectorListener = new ResponseStreamingWithoutBufferingListener();
+            TestUtil.updateMessageProcessor(httpConnectorListener);
+            HttpURLConnection urlConn = TestUtil.request(baseURI, "/", HttpMethod.POST.name(), true);
+            urlConn.setChunkedStreamingMode(-1); // Enable Chunking
+            TestUtil.writeContent(urlConn, requestValue);
+            assertEquals(200, urlConn.getResponseCode());
+            String content = TestUtil.getContent(urlConn);
+            assertEquals(requestValue, content);
+            urlConn.disconnect();
+        } catch (IOException e) {
+            TestUtil.handleException("IOException occurred while running responseStreamingWithoutBufferingTestCase", e);
+        }
+    }
 
     @AfterClass
     public void cleanUp() throws ServerConnectorException {
