+++ http/org.wso2.carbon.transport.http.netty/src/test/java/org/wso2/carbon/transport/http/netty/contentaware/listeners/RequestResponseTransformListener.java
@@ -17,7 +17,7 @@
  *
  */
 
-package org.wso2.carbon.transport.http.netty.contentaware.serverConnectorListeners;
+package org.wso2.carbon.transport.http.netty.contentaware.listeners;
 
 import io.netty.buffer.Unpooled;
 import io.netty.handler.codec.http.DefaultLastHttpContent;
@@ -48,16 +48,18 @@ import java.util.concurrent.Executors;
 import java.util.stream.Collectors;
 
 /**
- * A Message Processor which creates Request and Response
+ * Transform message in request and response path
  */
-public class RequestResponseCreationListener implements HttpConnectorListener {
-    private Logger logger = LoggerFactory.getLogger(RequestResponseCreationListener.class);
+public class RequestResponseTransformListener implements HttpConnectorListener {
 
-    private String responseValue;
-    private TransportsConfiguration configuration;
+    private static final Logger logger = LoggerFactory.getLogger(RequestResponseTransformListener.class);
+
     private ExecutorService executor = Executors.newSingleThreadExecutor();
+    private String responseValue;
+    private String requestValue;
+    private TransportsConfiguration configuration;
 
-    public RequestResponseCreationListener(String responseValue, TransportsConfiguration configuration) {
+    public RequestResponseTransformListener(String responseValue, TransportsConfiguration configuration) {
         this.responseValue = responseValue;
         this.configuration = configuration;
     }
@@ -68,21 +70,22 @@ public class RequestResponseCreationListener implements HttpConnectorListener {
             try {
                 int length = httpRequest.getFullMessageLength();
                 List<ByteBuffer> byteBufferList = httpRequest.getFullMessageBody();
-                ByteBuffer byteBuffer = ByteBuffer.allocate(length);
-                byteBufferList.forEach(byteBuffer::put);
-                String requestValue = new String(byteBuffer.array());
-                byte[] arry = responseValue.getBytes("UTF-8");
 
-                HTTPCarbonMessage newMsg = httpRequest.cloneCarbonMessageWithOutData();
-                if (newMsg.getHeader(Constants.HTTP_TRANSFER_ENCODING) == null) {
-                    newMsg.setHeader(Constants.HTTP_CONTENT_LENGTH, String.valueOf(arry.length));
+                ByteBuffer byteBuff = ByteBuffer.allocate(length);
+                byteBufferList.forEach(byteBuff::put);
+                requestValue = new String(byteBuff.array());
+
+                httpRequest.setProperty(Constants.HOST, TestUtil.TEST_HOST);
+                httpRequest.setProperty(Constants.PORT, TestUtil.TEST_HTTP_SERVER_PORT);
+
+                if (responseValue != null) {
+                    byte[] array = responseValue.getBytes("UTF-8");
+                    ByteBuffer byteBuffer = ByteBuffer.allocate(array.length);
+                    byteBuffer.put(array);
+                    httpRequest.setHeader(Constants.HTTP_CONTENT_LENGTH, String.valueOf(array.length));
+                    byteBuffer.flip();
+                    httpRequest.addHttpContent(new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuffer)));
                 }
-                ByteBuffer byteBuffer1 = ByteBuffer.allocate(arry.length);
-                byteBuffer1.put(arry);
-                byteBuffer1.flip();
-                newMsg.addHttpContent(new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuffer1)));
-                newMsg.setProperty(Constants.HOST, TestUtil.TEST_HOST);
-                newMsg.setProperty(Constants.PORT, TestUtil.TEST_HTTP_SERVER_PORT);
 
                 Map<String, Object> transportProperties = new HashMap<>();
                 Set<TransportProperty> transportPropertiesSet = configuration.getTransportProperties();
@@ -99,41 +102,35 @@ public class RequestResponseCreationListener implements HttpConnectorListener {
                 HttpWsConnectorFactory httpWsConnectorFactory = new HttpWsConnectorFactoryImpl();
                 HttpClientConnector clientConnector =
                         httpWsConnectorFactory.createHttpClientConnector(transportProperties, senderConfiguration);
-
-                HttpResponseFuture future = clientConnector.send(newMsg);
+                HttpResponseFuture future = clientConnector.send(httpRequest);
                 future.setHttpConnectorListener(new HttpConnectorListener() {
                     @Override
-                    public void onMessage(HTTPCarbonMessage httpResponse) {
+                    public void onMessage(HTTPCarbonMessage httpMessage) {
                         executor.execute(() -> {
-                            int length = httpResponse.getFullMessageLength();
-                            List<ByteBuffer> byteBufferList = httpResponse.getFullMessageBody();
+                            int length = httpMessage.getFullMessageLength();
+                            List<ByteBuffer> byteBufferList = httpMessage.getFullMessageBody();
 
                             ByteBuffer byteBuffer = ByteBuffer.allocate(length);
                             byteBufferList.forEach(byteBuffer::put);
                             String responseValue = new String(byteBuffer.array()) + ":" + requestValue;
-                            byte[] array = new byte[0];
-                            try {
-                                array = responseValue.getBytes("UTF-8");
-                            } catch (UnsupportedEncodingException e) {
+                            if (requestValue != null) {
+                                byte[] array = new byte[0];
+                                try {
+                                    array = responseValue.getBytes("UTF-8");
+                                } catch (UnsupportedEncodingException e) {
 
-                            }
-
-                            ByteBuffer byteBuff = ByteBuffer.allocate(array.length);
-                            byteBuff.put(array);
-                            byteBuff.flip();
-                            HTTPCarbonMessage httpCarbonMessage = httpResponse
-                                    .cloneCarbonMessageWithOutData();
-                            if (httpCarbonMessage.getHeader(Constants.HTTP_TRANSFER_ENCODING) == null) {
-                                httpCarbonMessage.setHeader(Constants.HTTP_CONTENT_LENGTH,
-                                        String.valueOf(array.length));
-                            }
-                            httpCarbonMessage.addHttpContent(
-                                    new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuff)));
-
-                            try {
-                                httpRequest.respond(httpCarbonMessage);
-                            } catch (ServerConnectorException e) {
-                                logger.error("Error occurred during message notification: " + e.getMessage());
+                                }
+                                ByteBuffer byteBuff = ByteBuffer.allocate(array.length);
+                                byteBuff.put(array);
+                                httpMessage.setHeader(Constants.HTTP_CONTENT_LENGTH, String.valueOf(array.length));
+                                byteBuff.flip();
+                                httpMessage.addHttpContent(
+                                        new DefaultLastHttpContent(Unpooled.wrappedBuffer(byteBuff)));
+                                try {
+                                    httpRequest.respond(httpMessage);
+                                } catch (ServerConnectorException e) {
+                                    logger.error("Error occurred during message notification: " + e.getMessage());
+                                }
                             }
                         });
                     }
@@ -143,13 +140,10 @@ public class RequestResponseCreationListener implements HttpConnectorListener {
 
                     }
                 });
-            } catch (UnsupportedEncodingException e) {
-                logger.error("Encoding is not supported", e);
             } catch (Exception e) {
-                logger.error("Failed to send the message to the back-end", e);
+                logger.error("Error while reading stream", e);
             }
         });
-
     }
 
     @Override
